requirements.txt

numpy==1.19.5
opencv-python==4.5.5.62
Pillow==8.4.0
streamlit==1.5.0
requests
requests-toolbelt


